complete update

+++ utils/augclass.py
@@ -196,7 +196,7 @@ class TimeReverse:
         # labels:0 for not reversed, 1 for reversed
         labels = torch.zeros(x.shape[0], dtype=torch.long)
         labels[applied_batch_idx] = 1
-        labels = torch.nn.functional.one_hot(labels, num_classes=2)
+        labels = torch.nn.functional.one_hot(labels, num_classes=2).float()
         if self.random:
             return x, labels
         else:
@@ -216,7 +216,7 @@ class SignFlip:
         labels = torch.zeros(x.shape[0], dtype=torch.long)
         labels[applied_batch_idx] = 1
         # one-hot encoding
-        labels = torch.nn.functional.one_hot(labels, num_classes=2)
+        labels = torch.nn.functional.one_hot(labels, num_classes=2).float()
         if self.random:
             return x, labels
         else:
